requirements.txt
requests
beautifulsoup4
transformers
torch
chromadb
pydantic
biopython
fitz
frontend
pymupdf
Retriever
farm-haystack[inference]
openai==0.28 
